Login Register i Refresh

+++ src/main/java/Studentski/sustav/sustav_backend/repositories/RoleRepository.java
@@ -1,4 +1,3 @@
-// java
 package Studentski.sustav.sustav_backend.repositories;
 
 import Studentski.sustav.sustav_backend.models.Role;
@@ -6,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
     Optional<Role> findByName(String name);
 }
